[fix] Honor `global-treesit-auto-modes`
Fixes #35

In `treesit-auto--set-major-remap`, bail if the major mode we are switching to is not part of `global-treesit-auto-modes`. The latter is a variable that was already declared as part of the call to `define-globalized-minor-mode`, but is not in use right now (at least not for the purpose that its doc string documents).

Here is how to configure treesit-auto to apply to bash only:

```elisp
(use-package treesit-auto
  :config
  (setq treesit-auto-langs '(bash)
        treesit-auto-install t
        global-treesit-auto-modes '(sh-mode))
  (global-treesit-auto-mode))
```

This change
- snarfs the implementation of `easy-mmode--globalized-predicate-p` and rewrites it into a pure function `treesit-auto--enabled-on-major-mode-p`
- calls `treesit-auto--enabled-on-major-mode-p` as part of the `set-auto-mode-0` advice, and avoids doing anything in case the result is nil.