Add extensions to treesit-auto recipes and register auto-mode-alist
In many cases, the major modes required for activating a tree-sitter
mode via `global-treesit-auto-mode` are not available out-of-the-box
with Emacs. This means that enabling a tree-sitter mode requires (a)
installing the parser and (b) installing the prerequisite major mode
before you can use the tree-sitter mode.

This commit adds a new function that registers tree-sitter modes based
on file extension, enabling use of tree-sitter major modes without
needing to install any other packages.