Remove the language from `treesit-auto-langs`
This was the underlying issue for https://github.com/renzmann/treesit-auto/pull/81.

In the README it states:

```md
By activating global-treesit-auto-mode, Emacs will:
• Automatically switch to <name>-ts-mode when the grammar for <name> is installed
• Stick with <name>-mode if the grammar isn’t installed
```

But if it can not switch to it, it kept on retrying, which caused an infinite loop unitl Emacs itself decided to kill the loop.

After this change, the recipe for Magik can be re-added again.

Sorry for the constant adding/removing of the recipe