make c and cpp recipes depend on each other
When `c-ts-mode.el` loaded, it evaluates

```lisp
(and (treesit-ready-p 'cpp)
     (treesit-ready-p 'c))
```

which pops up a warning every time even though treesit-auto install
mode specific tree-sitter. This was unexpected as we asked
treesit-auto to install libraries and it is okay with it. Users may
still get the warning but before libraries are installed and only once
across Emacs sessions.